+++ electron.vite.config.ts
@@ -25,6 +25,14 @@ export default defineConfig({
       alias: {
         '@images': resolve(__dirname, 'images')
       }
+    },
+    build: {
+      rollupOptions: {
+        input: {
+          index: resolve(__dirname, 'src/renderer/index.html'),
+          splash: resolve(__dirname, 'src/renderer/splash.html')
+        }
+      }
     }
   }
 });